+++ integration-tests/http/__tests__/order/admin/transfer-flow.spec.ts
@@ -59,6 +59,7 @@ medusaIntegrationTestRunner({
           `/admin/orders/${order.id}/transfer`,
           {
             customer_id: customer.id,
+            update_order_email: true,
           },
           adminHeaders
         )
@@ -112,6 +113,7 @@ medusaIntegrationTestRunner({
                 details: expect.objectContaining({
                   token: expect.any(String),
                   original_email: "tony@stark-industries.com",
+                  new_email: "test@email.com",
                 }),
               }),
             ]),
@@ -136,8 +138,8 @@ medusaIntegrationTestRunner({
           )
         ).data.order
 
-        expect(finalOrderResult.email).toEqual("tony@stark-industries.com")
-        // 4. Customer account is now associated with the order (email on the order is still as original, guest email)
+        // 4. Customer account is now associated with the order and email is updated to the customer's email
+        expect(finalOrderResult.email).toEqual("test@email.com")
         expect(finalOrderResult.customer_id).toEqual(customer.id)
       })
 
@@ -368,10 +370,10 @@ medusaIntegrationTestRunner({
                 action: "TRANSFER_CUSTOMER",
                 reference: "customer",
                 reference_id: customer.id,
-                details: expect.objectContaining({
+                details: {
                   token: expect.any(String),
                   original_email: "tony@stark-industries.com",
-                }),
+                },
               }),
             ]),
           })
@@ -386,8 +388,8 @@ medusaIntegrationTestRunner({
           )
         ).data.order
 
+        // 4. Customer account is now associated with the order and email is still the original email, since update_order_email is not true on the request
         expect(finalOrder.email).toEqual("tony@stark-industries.com")
-        // 4. Customer account is now associated with the order (email on the order is still as original, guest email)
         expect(finalOrder.customer_id).toEqual(customer.id)
       })