+++ packages/cli/medusa-cli/src/create-cli.ts
@@ -593,7 +593,7 @@ function isLocalMedusaProject() {
       `./package.json`
     ))
     // Draft order plugin can't have @medusajs/medusa as dependency,
-    // so we also check for @medusajs/cli 
+    // so we also check for @medusajs/cli
     inMedusaProject = !!(
       (dependencies &&
         (dependencies["@medusajs/medusa"] || dependencies["@medusajs/cli"])) ||